Event Consumer condtion

+++ event_streaming/event_streaming/doctype/event_update_log/event_update_log.py
@@ -17,7 +17,7 @@ class EventUpdateLog(Document):
 		jobs = get_jobs()
 		if not jobs or enqueued_method not in jobs[frappe.local.site]:
 			frappe.enqueue(
-				enqueued_method, doctype=self.ref_doctype, queue="long", enqueue_after_commit=True
+				enqueued_method, doctype=self.ref_doctype, queue="short", enqueue_after_commit=True
 			)
